🧹 Refactor and clean up analytics components and configurations
- Updated `.gitignore` to track additional lock files and exclude temporary files.
- Removed obsolete `AnalyticsProvider` component and related analytics tracking components to streamline the codebase.
- Refactored `layout.tsx` to use the new `DocsAnalyticsProvider` and updated font imports.
- Adjusted `package.json` to enforce `pnpm` as the package manager and added preinstall script.
- Cleaned up imports in `index.ts` for interactive components to follow consistent naming conventions.

These changes enhance the maintainability and clarity of the analytics implementation in the documentation site.
