+++ .npmrc
@@ -1,16 +1,1 @@
-# Enforce pnpm usage
-engine-strict=true
-
-# Prevent npm usage
-preinstall=npx only-allow pnpm
-
-# pnpm configuration
-shamefully-hoist=false
-strict-peer-dependencies=false
-auto-install-peers=true
-
-# Registry settings
 registry=https://registry.npmjs.org/
-
-# Node linker
-node-linker=isolated
