+++ apps/docs/src/components/interactive/analytics-dashboard.tsx
