+++ apps/docs/src/components/interactive/code-generator.tsx
