+++ apps/docs/src/components/interactive/index.ts
@@ -1,5 +1,5 @@
-export { ApiVisualizer } from './ApiVisualizer'
-export { DataFlowVisualization } from './DataFlowVisualization'
-export { CodeGenerator } from './CodeGenerator'
-export { AnalyticsDashboard } from './AnalyticsDashboard'
-export { RealTimeAnalyticsDashboard } from './RealTimeAnalyticsDashboard'
+export { ApiVisualizer } from './api-visualizer'
+export { DataFlowVisualization } from './data-flow-visualization'
+export { CodeGenerator } from './code-generator'
+export { AnalyticsDashboard } from './analytics-dashboard'
+export { RealTimeAnalyticsDashboard } from './real-time-analytics-dashboard'
