+++ package.json
@@ -2,11 +2,13 @@
   "name": "hono-analytics",
   "private": true,
   "type": "module",
+  "packageManager": "pnpm@9",
   "workspaces": [
     "apps/*",
     "packages/*"
   ],
   "scripts": {
+    "preinstall": "npx only-allow pnpm",
     "dev": "./dev.sh",
     "build": "pnpm -r build",
     "type-check": "pnpm -r type-check",
