+++ packages/backend/src/config/index.ts
@@ -1,4 +1,4 @@
-export { env, isDevelopment, isProduction } from './env.js'
+export { env } from './env.js'
 export { db, databaseConfig, schema } from './database.js'
 export { serverConfig, getApiInfo, getDatabaseInfo } from './server.js'
 export type { TDatabase } from './database.js'
